requirements.txt
alabaster==0.7.6
algorithms==0.1
altgraph==0.10.2
arrow==0.6.0
awsebcli==3.6.2
Babel==2.1.1
backports.ssl-match-hostname==3.4.0.2
bdist-mpkg==0.5.0
binaryornot==0.3.0
blessed==1.9.5
bonjour-py==0.3
botocore==1.3.14
cement==2.4.0
certifi==2015.4.28
click==3.3
colorama==0.3.3
cookiecutter==1.0.0
Django==1.8.5
django-bootstrap3==6.2.2
django-cors-headers==1.1.0
django-debug-toolbar==1.3.2
django-extensions==1.5.7
django-nose==1.4.2
djangorestframework==3.2.4
docker-compose==1.4.2
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
Factory==1.2
factory-boy==2.5.2
fake-factory==0.5.3
Faker==0.0.4
funcsigs==0.4
functools32==3.2.3.post2
gnureadline==6.3.3
httplib2==0.9.2
Inflector==2.0.11
ipaddress==1.0.14
ipython==3.2.1
Jinja2==2.7.3
jmespath==0.9.0
jsonfield==1.0.3
jsonschema==2.5.1
macholib==1.5.1
MarkupSafe==0.23
matplotlib==1.3.1
mistune==0.7
mock==1.3.0
models==0.9.3
modulegraph==0.10.4
nose==1.3.7
numpy==1.8.0rc1
pandas==0.16.2
pathspec==0.3.3
pbr==1.3.0
psycopg2==2.6.1
ptyprocess==0.5
py2app==0.7.3
pycountry==1.14
Pygments==2.0.2
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==2.4.2
pytz==2015.6
PyYAML==3.11
pyzmq==14.7.0
querystring-parser==1.2.1
requests==2.6.2
schedule==0.3.2
scipy==0.13.0b1
six==1.10.0
slate==0.3
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
sqlparse==0.1.16
statistics==1.0.3.5
stevedore==1.6.0
swagger-client==1.0.0
terminado==0.5
texttable==0.8.4
tornado==4.2.1
twilio==4.6.0
Unidecode==0.4.18
urbanairship==0.7.4
urllib3==1.11
utils==0.8.0
vboxapi==1.0
virtualenv==13.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.6.0
wcwidth==0.1.5
websocket-client==0.34.0
xattr==0.6.4
yolk==0.4.3
zope.interface==4.1.1
